backend/requirements/dev.txt
-r base.txt
django-extensions>=3.2
django-debug-toolbar>=4.2
ipython>=8.0
